+++ include/bitcoin/server/service/fetch_x.hpp
@@ -17,17 +17,32 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-#ifndef LIBBITCOIN_SERVER_COMPAT_HPP
-#define LIBBITCOIN_SERVER_COMPAT_HPP
+#ifndef LIBBITCOIN_SERVER_FETCH_X_HPP
+#define LIBBITCOIN_SERVER_FETCH_X_HPP
 
-#include "util.hpp"
+#include <bitcoin/server/node_impl.hpp>
+#include <bitcoin/server/service/util.hpp>
 
 namespace libbitcoin {
 namespace server {
 
-class node_impl;
+// fetch_history stuff
 
-void COMPAT_fetch_history(node_impl& node,
+bool unwrap_fetch_history_args(
+    bc::payment_address& payaddr, uint32_t& from_height,
+    const incoming_message& request);
+
+void send_history_result(
+    const std::error_code& ec, const bc::chain::history_list& history,
+    const incoming_message& request, queue_send_callback queue_send);
+
+// fetch_transaction stuff
+
+bool unwrap_fetch_transaction_args(
+    bc::hash_digest& tx_hash, const incoming_message& request);
+
+void transaction_fetched(const std::error_code& ec,
+    const bc::transaction_type& tx,
     const incoming_message& request, queue_send_callback queue_send);
 
 } // namespace server
