+++ include/bitcoin/server/version.hpp
@@ -7,6 +7,14 @@
 #ifndef LIBBITCOIN_SERVER_VERSION_HPP
 #define LIBBITCOIN_SERVER_VERSION_HPP
 
-#define LIBBITCOIN_SERVER_VERSION "2.0.0"
+/**
+ * The semantic version of this repository as: [major].[minor].[patch]
+ * For interpretation of the versioning scheme see: http://semver.org
+ */
+
+#define LIBBITCOIN_SERVER_VERSION "2.1.0"
+#define LIBBITCOIN_SERVER_MAJOR_VERSION 2
+#define LIBBITCOIN_SERVER_MINOR_VERSION 1
+#define LIBBITCOIN_SERVER_PATCH_VERSION 0
 
 #endif
