+++ src/message.cpp
@@ -17,8 +17,9 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
+#include <bitcoin/server/message.hpp>
+
 #include <random>
-#include "message.hpp"
 
 namespace libbitcoin {
 namespace server {
