+++ src/service/blockchain.cpp
@@ -17,12 +17,12 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-#include "blockchain.hpp"
+#include <bitcoin/server/service/blockchain.hpp>
 
-#include "../config.hpp"
-#include "../node_impl.hpp"
-#include "fetch_x.hpp"
-#include "util.hpp"
+#include <bitcoin/server/config.hpp>
+#include <bitcoin/server/node_impl.hpp>
+#include <bitcoin/server/service/fetch_x.hpp>
+#include <bitcoin/server/service/util.hpp>
 
 namespace libbitcoin {
 namespace server {
