+++ src/service/fullnode.cpp
@@ -17,11 +17,11 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-#include "fetch_x.hpp"
+#include <bitcoin/server/service/fullnode.hpp>
 
-#include "../config.hpp"
-#include "../node_impl.hpp"
-#include "fullnode.hpp"
+#include <bitcoin/server/config.hpp>
+#include <bitcoin/server/node_impl.hpp>
+#include <bitcoin/server/service/fetch_x.hpp>
 
 namespace libbitcoin {
 namespace server {
