+++ src/service/protocol.cpp
@@ -17,11 +17,11 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-#include "protocol.hpp"
+#include <bitcoin/server/service/protocol.hpp>
 
-#include "../config.hpp"
-#include "../node_impl.hpp"
-#include "util.hpp"
+#include <bitcoin/server/config.hpp>
+#include <bitcoin/server/node_impl.hpp>
+#include <bitcoin/server/service/util.hpp>
 
 namespace libbitcoin {
 namespace server {
