--- src/worker-output.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-watch -n 0.5 "grep worker debug.log | tail -n 20"
-
